+++ cartographer/io/points_processor_pipeline_builder.cc
@@ -19,7 +19,7 @@
 #include "absl/memory/memory.h"
 #include "cartographer/io/coloring_points_processor.h"
 #include "cartographer/io/counting_points_processor.h"
-#include "cartographer/io/dynamic_object_removal_points_processor.h"
+#include "cartographer/io/dynamic_objects_removal_points_processor.h"
 #include "cartographer/io/fixed_ratio_sampling_points_processor.h"
 #include "cartographer/io/frame_id_filtering_points_processor.h"
 #include "cartographer/io/hybrid_grid_points_processor.h"
@@ -83,7 +83,6 @@ void RegisterBuiltInPointsProcessors(
     const std::vector<mapping::proto::Trajectory>& trajectories,
     const FileWriterFactory& file_writer_factory,
     PointsProcessorPipelineBuilder* builder) {
-  RegisterFileWritingPointsProcessor<DynamicObjectsRemovalPointsProcessor>(file_writer_factory, builder);
   RegisterPlainPointsProcessor<CountingPointsProcessor>(builder);
   RegisterPlainPointsProcessor<FixedRatioSamplingPointsProcessor>(builder);
   RegisterPlainPointsProcessor<FrameIdFilteringPointsProcessor>(builder);
@@ -92,6 +91,7 @@ void RegisterBuiltInPointsProcessors(
   RegisterPlainPointsProcessor<OutlierRemovingPointsProcessor>(builder);
   RegisterPlainPointsProcessor<ColoringPointsProcessor>(builder);
   RegisterPlainPointsProcessor<IntensityToColorPointsProcessor>(builder);
+  RegisterPlainPointsProcessor<DynamicObjectsRemovalPointsProcessor>(builder);
   RegisterFileWritingPointsProcessor<PcdWritingPointsProcessor>(
       file_writer_factory, builder);
   RegisterFileWritingPointsProcessor<MeshWritingPointsProcessor>(
